passing core unit tests
