+++ core/services/mongodb-remote/src/main/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/DocumentVersionInfo.java
@@ -60,7 +60,7 @@ public final class DocumentVersionInfo {
       );
     }
 
-    private Version(
+    public Version(
             final int syncProtocolVersion,
             final String instanceId,
             final long versionCounter) {
