--- core/services/mongodb-remote/src/test/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/BaseChangeEventUnitTests.kt
@@ -1,2 +1,0 @@
-package com.mongodb.stitch.core.services.mongodb.remote.sync.internal
-
