+++ server/services/mongodb-remote/src/main/java/com/mongodb/stitch/server/services/mongodb/remote/internal/RemoteMongoCollectionImpl.java
@@ -523,7 +523,7 @@ public final class RemoteMongoCollectionImpl<DocumentT>
           final Class<ResultT> resultClass) {
     return proxy.findOneAndDelete(filter, options, resultClass);
   }
-  
+
   @Override
   public ChangeStream<ChangeEvent<DocumentT>, ChangeEvent<DocumentT>> watch(final ObjectId... ids)
       throws InterruptedException, IOException {
